Refactor NPU vLLM Ascend runtime patches
